+++ src/Component/Middleware/TenantIdMiddleware.php
@@ -40,11 +40,19 @@ class TenantIdMiddleware extends AbstractMiddleware
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
-        $this->constManager = $this->services->get('const_builder');
+        $this->constManager  = $this->services->get('const_builder');
         $this->isMultitenant = self::isMultitenantApp($this->configs->config['composer']);
-        $this->kioskConfig = $this->configs->config['kiosk'];
-        $this->tenantDomain = $this->resolveTenantIdFromRequest($request);
-        $this->isAdminKiosk = $this->isKiosk($this->tenantDomain);
+        $this->kioskConfig   = $this->configs->config['kiosk'];
+        $this->tenantDomain  = $this->resolveTenantIdFromRequest($request);
+        $this->isAdminKiosk  = $this->isKiosk($this->tenantDomain);
+        $this->isShortInit   = $this->isShortInit();
+
+        // Add attributes to the request
+        $request = $request
+            ->withAttribute('isAdminKiosk', $this->isAdminKiosk)
+            ->withAttribute('isMultitenant', $this->isMultitenant)
+            ->withAttribute('isShortInit', $this->isShortInit)
+            ->withAttribute('tenant', $this->tenant);
 
         // check secure modes.
         if (self::isSecureMode() && 'https' !== $request->getUri()->getScheme()) {
@@ -59,13 +67,10 @@ class TenantIdMiddleware extends AbstractMiddleware
 
         // Set the current tenant
         $this->tenant = $this->setCurrentTenant();
+        $request = $request->withAttribute('tenant', $this->tenant);
 
         // @phpstan-ignore-next-line
         if ($this->isAdminKiosk && $this->kioskConfig->get('panel.enabled', null)) {
-            $request = $request
-                ->withAttribute('isAdminKiosk', $this->isAdminKiosk)
-                ->withAttribute('tenant', $this->tenant);
-
             return $handler->handle($request);
         }
 
@@ -75,11 +80,6 @@ class TenantIdMiddleware extends AbstractMiddleware
 
             throw new Exception("Tenant {$this->tenantDomain[0]}: {$tenantStatus}", 404);
         }
-
-        // Add attributes to the request for multitenant applications
-        $request = $request
-            ->withAttribute('tenant', $this->tenant)
-            ->withAttribute('isMultitenant', $this->isMultitenant);
 
         // Define required constants
         \define('APP_TENANT_ID', $this->tenant['uuid']);
@@ -144,8 +144,12 @@ class TenantIdMiddleware extends AbstractMiddleware
         return $this->tenant;
     }
 
-    protected function isKiosk(array $tenantDomain): bool
+    protected function isKiosk(array $tenantDomain): ?bool
     {
+        if (empty($tenantDomain)) {
+            return false;
+        }
+
         $kioskDomain = env('KIOSK_DOMAIN_ID', $this->kioskConfig->get('panel.id', 'kiosk'));
 
         return $kioskDomain === ($tenantDomain[0] ?? null);
@@ -193,7 +197,7 @@ class TenantIdMiddleware extends AbstractMiddleware
      *
      * @psalm-return list{string, string}|null
      */
-    private function resolveTenantIdFromRequest(ServerRequestInterface $request): ?array
+    private function resolveTenantIdFromRequest(ServerRequestInterface $request): array
     {
         $host = $request->getUri()->getHost();
         $domainOrSubdomain = explode('.', $host)[0];
@@ -202,7 +206,12 @@ class TenantIdMiddleware extends AbstractMiddleware
             return [$domainOrSubdomain,$host];
         }
 
-        return null;
+        return [];
+    }
+
+    private function isShortInit(): bool
+    {
+        return \defined('SHORTINIT') && true === \constant('SHORTINIT');
     }
 
     /**